Consolidate duplicate controllers, clean up routes
Issue #869
