+++ app/Http/Controllers/GroupController.php
@@ -13,7 +13,7 @@ use App\Models\Group;
 use App\Models\GroupMember;
 use App\Models\MadisonEvent;
 
-class GroupsController extends Controller
+class GroupController extends Controller
 {
     public function __construct()
     {
@@ -374,4 +374,54 @@ class GroupsController extends Controller
 
         return Redirect::to('groups')->with('success_message', $message);
     }
+
+    public function getVerify()
+    {
+        $this->beforeFilter('admin');
+
+        $status = Input::get('status');
+
+        if ($status) {
+            $groups = Group::where('status', '=', $status)->get();
+        } else {
+            $groups = Group::all();
+        }
+
+        return Response::json($groups);
+    }
+
+    public function putVerify($groupId)
+    {
+        $this->beforeFilter('admin');
+
+        $newGroup = (object) Input::all();
+
+        if (!Group::isValidStatus($newGroup->status)) {
+            throw new \Exception("Invalid value for verify request");
+        }
+
+        $group = Group::where('id', '=', $groupId)->first();
+
+        if (!$group) {
+            throw new \Exception("Invalid Group");
+        }
+
+        $group->status = $newGroup->status;
+
+        DB::transaction(function () use ($group) {
+            $group->save();
+
+            switch ($group->status) {
+                case Group::STATUS_ACTIVE:
+                    $group->createRbacRules();
+                    break;
+                case Group::STATUS_PENDING:
+                    $group->destroyRbacRules();
+                    break;
+            }
+        });
+
+        return Response::json($group);
+    }
+
 }
