+++ app/Http/routes.php
@@ -66,9 +66,9 @@ Route::get('sitemap', function () {
 |
 */
 
-/*
-*   Global Route Patterns
-*/
+/**
+ * Global Route Patterns
+ */
 
 Route::pattern('annotation', '[0-9a-zA-Z_-]+');
 Route::pattern('comment', '[0-9a-zA-Z_-]+');
@@ -79,17 +79,11 @@ Route::pattern('group', '[0-9]+');
 Route::pattern('image', '[a-zA-Z0-9-_]+\.[a-zA-Z0-9]{2,4}');
 Route::pattern('state', Doc::validStatesPattern());
 
-/*
-*   Route - Model bindings
-*/
+/**
+ * Route - Model bindings
+ */
 Route::model('user', 'App\Models\User');
 Route::model('user/edit', 'App\Models\User');
-
-/*
-|--------------------------------------------------------------------------
-| Application Routes
-|--------------------------------------------------------------------------
-*/
 
 // Modal Routes
 Route::get('modals/annotation_thanks', array(
@@ -113,167 +107,125 @@ Route::get('api/settings/vendors', function () {
     return ['uservoice' => $uservoice, 'ga' => $ga];
 });
 
-Route::post('groups/member/{memberId}/role', 'GroupsController@changeMemberRole');
-Route::post('api/groups/active/{groupId}', 'GroupsController@setActiveGroup');
+// Document Routes
+Route::get('api/docs/', 'DocumentController@getDocs');
+Route::post('api/docs/', 'DocumentController@postDocs');
+Route::get('api/docs/count', 'DocumentController@getDocCount');
+Route::get('api/docs/{state}', 'DocumentController@getDocs');
+Route::post('api/docs/featured', 'DocumentController@postFeatured');
+Route::put('api/docs/featured', 'DocumentController@putFeatured');
+Route::delete('api/docs/featured/{doc}', 'DocumentController@deleteFeatured');
+Route::get('api/docs/recent/{query?}', 'DocumentController@getRecent')->where('query', '[0-9]+');
+Route::get('api/docs/categories', 'DocumentController@getCategories');
+Route::get('api/docs/statuses', 'DocumentController@getAllStatuses');
+Route::get('api/docs/sponsors', 'DocumentController@getAllSponsors');
+Route::get('api/docs/featured', 'DocumentController@getFeatured');
+Route::get('api/docs/deleted', 'DocumentController@getDeletedDocs')->middleware(['auth']);
+Route::put('api/dates/{date}', 'DocumentController@putDate');
 
-Route::get('api/groups/roles', 'GroupsController@getRoles');
-Route::get('api/groups/{group?}', 'GroupsController@getGroup');
-Route::post('api/groups/{group?}', 'GroupsController@postGroup');
-Route::delete('api/groups/{groupId}/members/{memberId}', 'GroupsController@removeMember');
-Route::put('api/groups/{groupId}/invite', 'GroupsController@processMemberInvite');
-Route::get('api/groups/{groupId}/members', 'GroupsController@getMembers');
-Route::put('api/groups/{groupId}/members/{memberId}', 'GroupsController@putMember');
+// Single Doc Routes
+Route::post('api/docs/{doc}/support/', 'DocumentController@postSupport')->middleware(['doc.access.read']);
+Route::get('api/users/{user}/support/{doc}', 'UserController@getSupport')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/categories', 'DocumentController@getCategories')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/introtext', 'DocumentController@getIntroText')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/content', 'DocumentController@getContent')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/sponsor/{sponsor}', 'DocumentController@hasSponsor')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/sponsor', 'DocumentController@getSponsor')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/status', 'DocumentController@getStatus')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/dates', 'DocumentController@getDates')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/images/{image}','DocumentController@getImage')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}', 'DocumentController@getDoc')->middleware(['doc.access.read']);
+Route::get('api/docs/slug/{slug}', 'DocumentController@getDocBySlug')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/introtext', 'DocumentController@postIntroText')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/title', 'DocumentController@postTitle')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/sponsor', 'DocumentController@postSponsor')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/publishstate', 'DocumentController@postPublishState')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/slug', 'DocumentController@postSlug')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/content', 'DocumentController@postContent')->middleware(['doc.access.edit']);
+Route::put('api/docs/{doc}/content/{page}', 'DocumentController@putContent')->middleware(['doc.access.edit']);
+Route::delete('api/docs/{doc}/content/{page}', 'DocumentController@deleteContent')->middleware(['doc.access.edit']);
+Route::get('docs/embed/{slug}', 'DocumentController@getEmbedded')->middleware(['doc.access.read']);
+Route::get('api/docs/{slug}/feed', 'DocumentController@getFeed')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/featured-image', 'DocumentController@uploadImage')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/status', 'DocumentController@postStatus')->middleware(['doc.access.edit']);
+Route::delete('api/docs/{doc}', 'DocumentController@deleteDoc')->middleware(['doc.access.edit']);
+Route::put('api/docs/{doc}/restore', 'DocumentController@getRestoreDoc')->middleware(['doc.access.edit']);
+Route::delete('api/docs/{doc}/featured-image', 'DocumentController@deleteImage')->middleware(['doc.access.edit']);
+Route::delete('api/docs/{doc}/dates/{date}', 'DocumentController@deleteDate')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/dates', 'DocumentController@postDate')->middleware(['doc.access.edit']);
+Route::post('api/docs/{doc}/categories', 'DocumentController@postCategories')->middleware(['doc.access.edit']);
 
-//Document Routes
-Route::get('docs', 'DocController@index');
-Route::get('docs/{slug}', 'DocController@index')->middleware(['doc.access.read']);
-Route::get('docs/embed/{slug}', 'DocController@getEmbedded')->middleware(['doc.access.read']);
-Route::get('api/docs/{slug}/feed', 'DocController@getFeed')->middleware(['doc.access.read']);
-Route::get('documents/search', 'DocumentsController@getSearch');
-Route::get('documents', 'DocumentsController@listDocuments');
-Route::get('documents/view/{documentId}', 'DocumentsController@viewDocument');
-Route::get('documents/edit/{documentId}', 'DocumentsController@editDocument');
-Route::put('documents/edit/{documentId}', 'DocumentsController@saveDocumentEdits');
-Route::post('documents/create', 'DocumentsController@createDocument');
-Route::post('documents/save', 'DocumentsController@saveDocument');
-Route::delete('/documents/delete/{slug}', 'DocumentsController@deleteDocument')->middleware(['doc.access.edit']);
+// Annotation Action Routes
+Route::post('api/docs/{doc}/annotations/{annotation}/likes', 'AnnotationController@postLikes');
+Route::post('api/docs/{doc}/annotations/{annotation}/flags', 'AnnotationController@postFlags');
+Route::post('api/docs/{doc}/annotations/{annotation}/seen', 'AnnotationController@postSeen');
+Route::get('api/docs/{doc}/annotations/{annotation}/likes', 'AnnotationController@getLikes');
+Route::get('api/docs/{doc}/annotations/{annotation}/flags', 'AnnotationController@getFlags');
 
-//User Routes
-Route::get('user/{user}', 'UserController@getIndex');
-Route::get('user/edit/{user}', 'UserController@getEdit');
-Route::put('api/user/edit/{user}', 'UserController@putEdit');
-Route::post('api/user/verify-email', 'UserController@postVerify');
+// Annotation Comment Routes
+Route::get('api/docs/{doc}/annotations/{annotation}/comments', 'AnnotationController@getComments');
+Route::post('api/docs/{doc}/annotations/{annotation}/comments', 'AnnotationController@postComments');
+Route::get('api/docs/{doc}/annotations/{annotation}/comments/{comment}', 'AnnotationController@getComments');
 
-Route::post('api/password/remind', 'RemindersController@postRemind');
-Route::post('api/password/reset',  'RemindersController@postReset');
+// Annotation Routes
+Route::get('api/annotations/search', 'AnnotationController@getSearch');
+Route::get('api/docs/{doc}/annotations/{annotation?}', 'AnnotationController@getIndex')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/annotations', 'AnnotationController@postIndex');
+Route::put('api/docs/{doc}/annotations/{annotation}', 'AnnotationController@putIndex');
+Route::delete('api/docs/{doc}/annotations/{annotation}', 'AnnotationController@deleteIndex');
 
-// Confirmation email resend
-Route::post('api/verification/resend',  'RemindersController@postConfirmation');
+// Document Comment Routes
+Route::post('api/docs/{doc}/comments', 'CommentController@postIndex')->middleware(['doc.access.read']);
+Route::get('api/docs/{doc}/comments', 'CommentController@getIndex')->middleware(['doc.access.read']);;
+Route::get('api/docs/{doc}/comments/{comment?}', 'CommentController@getComment')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/comments/{comment}/likes', 'CommentController@postLikes')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/comments/{comment}/flags', 'CommentController@postFlags')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/comments/{comment}/comments', 'CommentController@postComments')->middleware(['doc.access.read']);
+Route::post('api/docs/{doc}/comments/{comment}/seen', 'CommentController@postSeen')->middleware(['doc.access.read']);
 
-//Annotation Routes
-Route::get('annotation/{annotation}', 'AnnotationController@getIndex');
-
-//Dashboard Routes
-Route::controller('dashboard', 'DashboardController');
-
-//Api Routes
-// Document API Routes
-Route::get('api/user/sponsors/all', 'DocumentApiController@getAllSponsorsForUser');
-Route::get('api/sponsors/all', 'SponsorApiController@getAllSponsors');
-
-//Annotation Action Routes
-Route::post('api/docs/{doc}/annotations/{annotation}/likes', 'AnnotationApiController@postLikes');
-Route::post('api/docs/{doc}/annotations/{annotation}/flags', 'AnnotationApiController@postFlags');
-Route::post('api/docs/{doc}/annotations/{annotation}/seen', 'AnnotationApiController@postSeen');
-Route::get('api/docs/{doc}/annotations/{annotation}/likes', 'AnnotationApiController@getLikes');
-Route::get('api/docs/{doc}/annotations/{annotation}/flags', 'AnnotationApiController@getFlags');
-
-//Annotation Comment Routes
-Route::get('api/docs/{doc}/annotations/{annotation}/comments', 'AnnotationApiController@getComments');
-Route::post('api/docs/{doc}/annotations/{annotation}/comments', 'AnnotationApiController@postComments');
-Route::get('api/docs/{doc}/annotations/{annotation}/comments/{comment}', 'AnnotationApiController@getComments');
-
-//Annotation Routes
-Route::get('api/annotations/search', 'AnnotationApiController@getSearch');
-Route::get('api/docs/{doc}/annotations/{annotation?}', 'AnnotationApiController@getIndex')->middleware(['doc.access.read']);
-Route::post('api/docs/{doc}/annotations', 'AnnotationApiController@postIndex');
-Route::put('api/docs/{doc}/annotations/{annotation}', 'AnnotationApiController@putIndex');
-Route::delete('api/docs/{doc}/annotations/{annotation}', 'AnnotationApiController@deleteIndex');
-
-//Document Routes
-Route::get('api/docs/slug/{slug}', 'DocumentsController@getDocument')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/content', 'DocumentsController@getDocumentContent')->middleware(['doc.access.read']);
-
-//Document Comment Routes
-Route::post('api/docs/{doc}/comments', 'CommentApiController@postIndex')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/comments', 'CommentApiController@getIndex')->middleware(['doc.access.read']);;
-Route::get('api/docs/{doc}/comments/{comment?}', 'CommentApiController@getComment')->middleware(['doc.access.read']);
-Route::post('api/docs/{doc}/comments/{comment}/likes', 'CommentApiController@postLikes')->middleware(['doc.access.read']);
-Route::post('api/docs/{doc}/comments/{comment}/flags', 'CommentApiController@postFlags')->middleware(['doc.access.read']);
-Route::post('api/docs/{doc}/comments/{comment}/comments', 'CommentApiController@postComments')->middleware(['doc.access.read']);
-Route::post('api/docs/{doc}/comments/{comment}/seen', 'CommentApiController@postSeen')->middleware(['doc.access.read']);
-
-//Document Support / Oppose routes
-Route::post('api/docs/{doc}/support/', 'DocController@postSupport')->middleware(['doc.access.read']);
-Route::get('api/users/{user}/support/{doc}', 'UserApiController@getSupport')->middleware(['doc.access.read']);
-
-//Document Api Routes
-Route::get('api/docs/recent/{query?}', 'DocumentApiController@getRecent')->where('query', '[0-9]+');
-Route::get('api/docs/active/{query?}', 'DocumentsController@getActive')->where('query', '[0-9]+');
-Route::get('api/docs/categories', 'DocumentApiController@getCategories');
-Route::get('api/docs/statuses', 'DocumentApiController@getAllStatuses');
-Route::get('api/docs/sponsors', 'DocumentApiController@getAllSponsors');
-Route::get('api/docs/featured', 'DocumentsController@getFeatured');
-Route::get('api/docs/deleted', 'DocumentApiController@getDeletedDocs')->middleware(['auth']);
-
-Route::get('api/docs/count', 'DocumentApiController@getDocCount');
-Route::get('api/docs/', 'DocumentApiController@getDocs');
-Route::get('api/docs/{state}', 'DocumentApiController@getDocs');
-Route::put('api/dates/{date}', 'DocumentApiController@putDate');
-
-Route::post('api/docs/featured', 'DocumentsController@postFeatured');
-Route::put('api/docs/featured', 'DocumentsController@putFeatured');
-Route::delete('api/docs/featured/{doc}', 'DocumentsController@deleteFeatured');
-
-Route::get('api/docs/{doc}/categories', 'DocumentApiController@getCategories')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/introtext', 'DocumentApiController@getIntroText')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/sponsor/{sponsor}', 'DocumentApiController@hasSponsor')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/sponsor', 'DocumentApiController@getSponsor')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/status', 'DocumentApiController@getStatus')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/dates', 'DocumentApiController@getDates')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}/images/{image}','DocumentsController@getImage')->middleware(['doc.access.read']);
-Route::get('api/docs/{doc}', 'DocumentApiController@getDoc')->middleware(['doc.access.read']);
-
-Route::post('api/docs/', 'DocumentApiController@postDocs');
-Route::post('api/docs/{doc}/introtext', 'DocumentApiController@postIntroText')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/title', 'DocumentApiController@postTitle')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/sponsor', 'DocumentApiController@postSponsor')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/publishstate', 'DocumentApiController@postPublishState')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/slug', 'DocumentApiController@postSlug')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/content', 'DocumentApiController@postContent')->middleware(['doc.access.edit']);
-Route::put('api/docs/{doc}/content/{page}', 'DocumentApiController@putContent')->middleware(['doc.access.edit']);
-Route::delete('api/docs/{doc}/content/{page}', 'DocumentApiController@deleteContent')->middleware(['doc.access.edit']);
-
-Route::post('api/docs/{doc}/featured-image', 'DocumentsController@uploadImage')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/status', 'DocumentApiController@postStatus')->middleware(['doc.access.edit']);
-Route::delete('api/docs/{doc}', 'DocumentApiController@deleteDoc')->middleware(['doc.access.edit']);
-Route::put('api/docs/{doc}/restore', 'DocumentApiController@getRestoreDoc')->middleware(['doc.access.edit']);
-Route::delete('api/docs/{doc}/featured-image', 'DocumentsController@deleteImage')->middleware(['doc.access.edit']);
-Route::delete('api/docs/{doc}/dates/{date}', 'DocumentApiController@deleteDate')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/dates', 'DocumentApiController@postDate')->middleware(['doc.access.edit']);
-Route::post('api/docs/{doc}/categories', 'DocumentApiController@postCategories')->middleware(['doc.access.edit']);
-
-
-//User Routes
-Route::get('api/user/{user}', 'UserApiController@getUser');
-Route::get('api/user/verify/', 'UserApiController@getVerify');
-Route::post('api/user/verify/', 'UserApiController@postVerify');
-Route::put('api/user/sponsor', 'SponsorApiController@putRequest');
-Route::get('api/user/admin/', 'UserApiController@getAdmins');
-Route::post('api/user/admin/', 'UserApiController@postAdmin');
-Route::get('api/user/independent/verify/', 'UserApiController@getIndependentVerify');
-Route::post('api/user/independent/verify/', 'UserApiController@postIndependentVerify');
+// User Routes
+Route::get('api/user/{user}', 'UserController@getUser')->middleware(['auth']);
+Route::get('api/user/verify/', 'UserController@getVerify')->middleware(['auth']);
+Route::post('api/user/verify/', 'UserController@postVerify')->middleware(['auth']);
+Route::put('api/user/sponsor', 'SponsorController@putRequest');
+Route::get('api/user/admin/', 'UserController@getAdmins')->middleware(['auth']);
+Route::post('api/user/admin/', 'UserController@postAdmin')->middleware(['auth']);
+Route::get('api/user/independent/verify/', 'UserController@getIndependentVerify')->middleware(['auth']);
+Route::post('api/user/independent/verify/', 'UserController@postIndependentVerify')->middleware(['auth']);
 Route::get('api/user/current', 'UserController@getCurrent');
 Route::put('api/user/{user}/edit/email', 'UserController@editEmail');
-Route::get('api/user/{user}/docs', 'DocumentsController@listDocuments');
+Route::get('api/user/{user}/docs', 'DocumentController@getUserDocuments');
 Route::get('api/user/{user}/notifications', 'UserController@getNotifications');
 Route::put('api/user/{user}/notifications', 'UserController@putNotifications');
 Route::get('api/user/{user}/groups', 'UserController@getGroups');
 Route::get('api/user/facebook-login', 'UserController@getFacebookLogin');
 Route::get('api/user/twitter-login', 'UserController@getTwitterLogin');
 Route::get('api/user/linkedin-login', 'UserController@getLinkedinLogin');
+Route::put('api/user/edit/{user}', 'UserController@putEdit');
+Route::post('api/user/verify-email', 'UserController@postVerifyEmail');
+Route::post('api/password/remind', 'RemindersController@postRemind');
+Route::post('api/password/reset',  'RemindersController@postReset');
+Route::get('api/user/sponsors/all', 'DocumentController@getAllSponsorsForUser');
+Route::get('api/sponsors/all', 'SponsorController@getAllSponsors');
+Route::post('api/verification/resend',  'RemindersController@postConfirmation');
 
 // Group Routes
-Route::get('api/groups/verify/', 'GroupsApiController@getVerify');
-Route::put('api/groups/verify/{groupId}', 'GroupsApiController@putVerify');
+Route::get('api/groups/verify/', 'GroupController@getVerify')->middleware(['auth']);
+Route::put('api/groups/verify/{groupId}', 'GroupController@putVerify')->middleware(['auth']);
+Route::post('api/groups/active/{groupId}', 'GroupController@setActiveGroup');
+Route::get('api/groups/roles', 'GroupController@getRoles');
+Route::get('api/groups/{group?}', 'GroupController@getGroup');
+Route::post('api/groups/{group?}', 'GroupController@postGroup');
+Route::delete('api/groups/{groupId}/members/{memberId}', 'GroupController@removeMember');
+Route::put('api/groups/{groupId}/invite', 'GroupController@processMemberInvite');
+Route::get('api/groups/{groupId}/members', 'GroupController@getMembers');
+Route::put('api/groups/{groupId}/members/{memberId}', 'GroupController@putMember');
 
 // User Login / Signup AJAX requests
-Route::get('api/user/login', 'UserManageApiController@getLogin');
-Route::post('api/user/login', 'UserManageApiController@postLogin');
-Route::get('api/user/signup', 'UserManageApiController@getSignup');
-Route::post('api/user/signup', 'UserManageApiController@postSignup');
+Route::post('api/user/login', 'UserController@postLogin');
+Route::post('api/user/signup', 'UserController@postSignup');
 
-//Auth Token Route
+// Auth Token Route
 Route::get('/auth/token', 'AuthController@token');
 Route::get('/api/user/login', 'AuthController@login');
 Route::get('/api/user/logout', 'AuthController@logout');
